+++ pkg/database/database.go
@@ -35,11 +35,8 @@ func Setup() {
 
 	if driver == "sqlite" {
 
+		// Default value is set on configuration.go
 		path := viper.GetString("database.path")
-		if len(path) == 0 {
-			log.Println("no database.path provided in config file for sqlite. using default:", path)
-			path = "./store/passwall.db"
-		}
 		db, err = gorm.Open("sqlite3", path)
 		if err != nil {
 			log.Fatal("db err: ", err)